feat(LottoController): LottoController생성하여 MVC 패턴 적용
